requirements.txt
numpy
pyserial
matplotlib